Use Cargo features to conditionally compile
This commit modifies the build process to enable certain dependencies
for different ROMs and/or CoDi firmware(s). This means the daemon can be
built specifically for one ROM and/or different CoDi types.

I've also added the new CoDi protocol crates hosted on the same organisation
as this repo.

TODO: Add check in `lib.rs` for at least *two* features to be enabled,
and fail the build if not available.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>